use negative -R to enable cachebox defined in the input file, fix time wraping

+++ mcextreme_cuda/trunk/src/tictoc.c
@@ -48,11 +48,11 @@ static unsigned int timerRes;
 #include <string.h>
 void SetupMillisTimer(void) {}
 void CleanupMillisTimer(void) {}
-int GetTime (void) {
+long GetTime (void) {
   struct timeval tv;
   timerRes = 1000;
   gettimeofday(&tv,NULL);
-  int temp = tv.tv_usec;
+  long temp = tv.tv_usec;
   temp+=tv.tv_sec*1000000;
   return temp;
 }
@@ -93,7 +93,7 @@ int GetTime(void)
       return 0;
    }
 
-   return ((int) (((double) counter.QuadPart) * cycles_per_usec));
+   return ((long) (((double) counter.QuadPart) * cycles_per_usec));
 }
 
 #pragma comment(lib,"winmm.lib")
